operator: Create separate metrics package
